// src/app/api/models/blood-type.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Schema to hold Blood Type information
 */
export interface BloodType {

  /**
   * Status
   */
  active?: boolean;

  /**
   * Blood Type Code
   */
  code: string;

  /**
   * Identificador único
   */
  id?: number;

  /**
   * Blood Type Name
   */
  name: string;
}